Fix datatype to accept numbers over one million.

+++ database/migrations/2021_03_11_222202_create_range_prices_table.php
@@ -16,8 +16,8 @@ class CreateRangePricesTable extends Migration
         Schema::create('range_prices', function (Blueprint $table) {
             $table->id();
             $table->string('value');
-            $table->float('big_price');
-            $table->float('small_price');
+            $table->decimal('big_price', 12, 2);
+            $table->decimal('small_price', 12, 2);
             $table->timestamps();
         });
     }
